+++ container-engine-lib/lib/backend_interface/objects/service/service_status.go
@@ -5,6 +5,7 @@ type UserServiceStatus int
 const (
 	UserServiceStatus_Registered UserServiceStatus = iota
 	UserServiceStatus_Running
+	UserServiceStatus_Paused	// Indicates that the service has been paused, and all processes frozen
 	UserServiceStatus_Stopped	// Indicates that the Service has been stopped and can no longer be used
 )